+++ pkg/sink/codec/canal/type_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/pingcap/tiflow/cdc/entry"
 	"github.com/pingcap/tiflow/cdc/model"
 	"github.com/pingcap/tiflow/pkg/sink/codec/internal"
+	"github.com/pingcap/tiflow/pkg/sink/codec/utils"
 	"github.com/stretchr/testify/require"
 )
 
@@ -40,10 +41,10 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[0].ID)
 	require.True(t, ok)
 
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "int", mysqlType)
 	// mysql type with the default type length
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "int(11)", mysqlType)
 
 	flag := tableInfo.ColumnsFlag[colInfos[0].ID]
@@ -53,9 +54,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinyint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinyint(4)", mysqlType)
 
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
@@ -65,9 +66,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "smallint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "smallint(6)", mysqlType)
 
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
@@ -78,9 +79,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "mediumint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "mediumint(9)", mysqlType)
 	javaType, err = getJavaSQLType(int64(8388607), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -89,9 +90,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "bigint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "bigint(20)", mysqlType)
 	javaType, err = getJavaSQLType(int64(9223372036854775807), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -111,10 +112,10 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[0].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[0].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "int unsigned", mysqlType)
 	// mysql type with the default type length
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "int(10) unsigned", mysqlType)
 
 	javaType, err = getJavaSQLType(uint64(2147483647), columnInfo.FieldType.GetType(), flag)
@@ -133,9 +134,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinyint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinyint(3) unsigned", mysqlType)
 
 	javaType, err = getJavaSQLType(uint64(127), columnInfo.FieldType.GetType(), flag)
@@ -154,9 +155,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "smallint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "smallint(5) unsigned", mysqlType)
 	javaType, err = getJavaSQLType(uint64(32767), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -174,9 +175,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "mediumint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "mediumint(8) unsigned", mysqlType)
 	javaType, err = getJavaSQLType(uint64(8388607), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -194,9 +195,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "bigint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "bigint(20) unsigned", mysqlType)
 	javaType, err = getJavaSQLType(uint64(9223372036854775807), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -225,41 +226,41 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[0].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[0].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "int", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "int(10)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinyint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinyint(3)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "smallint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "smallint(5)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "mediumint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "mediumint(8)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "bigint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "bigint(19)", mysqlType)
 
 	sql = `create table test.t4 (
@@ -276,41 +277,41 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[0].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[0].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "int unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "int(10) unsigned", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinyint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinyint(3) unsigned", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "smallint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "smallint(5) unsigned", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "mediumint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "mediumint(8) unsigned", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "bigint unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "bigint(19) unsigned", mysqlType)
 
 	sql = `create table test.t5 (
@@ -327,41 +328,41 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[0].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[0].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "int unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "int(10) unsigned zerofill", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinyint unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinyint(3) unsigned zerofill", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "smallint unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "smallint(5) unsigned zerofill", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "mediumint unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "mediumint(8) unsigned zerofill", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "bigint unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "bigint(20) unsigned zerofill", mysqlType)
 
 	sql = `create table test.t6(
@@ -377,17 +378,17 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "bit", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "bit(1)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "bit", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "bit(3)", mysqlType)
 	javaType, err = getJavaSQLType(uint64(65), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -397,9 +398,9 @@ func TestGetMySQLType4IntTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinyint", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinyint(1)", mysqlType)
 }
 
@@ -419,9 +420,9 @@ func TestGetMySQLType4FloatType(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "float", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "float", mysqlType)
 	javaType, err := getJavaSQLType(3.14, columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -430,9 +431,9 @@ func TestGetMySQLType4FloatType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "double", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "double", mysqlType)
 	javaType, err = getJavaSQLType(2.71, columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -447,17 +448,17 @@ func TestGetMySQLType4FloatType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "float", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "float(10,3)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "float", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "float", mysqlType)
 
 	sql = `create table test.t3(a int primary key, b double(20, 3))`
@@ -470,9 +471,9 @@ func TestGetMySQLType4FloatType(t *testing.T) {
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
 
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "double", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "double(20,3)", mysqlType)
 
 	sql = `create table test.t4(
@@ -489,9 +490,9 @@ func TestGetMySQLType4FloatType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "float unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "float unsigned", mysqlType)
 	javaType, err = getJavaSQLType(3.14, columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -500,9 +501,9 @@ func TestGetMySQLType4FloatType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "double unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "double unsigned", mysqlType)
 	javaType, err = getJavaSQLType(2.71, columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -511,17 +512,17 @@ func TestGetMySQLType4FloatType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "float unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "float unsigned zerofill", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "double unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "double unsigned zerofill", mysqlType)
 }
 
@@ -538,17 +539,17 @@ func TestGetMySQLType4Decimal(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "decimal", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "decimal(10,0)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "decimal", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "decimal(10,0)", mysqlType)
 
 	sql = `create table test.t2(a int primary key, b decimal(5), c decimal(5, 2))`
@@ -560,17 +561,17 @@ func TestGetMySQLType4Decimal(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "decimal", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "decimal(5,0)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "decimal", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "decimal(5,2)", mysqlType)
 	javaType, err := getJavaSQLType("2333", columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -585,17 +586,17 @@ func TestGetMySQLType4Decimal(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "decimal unsigned", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "decimal(10,0) unsigned", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "decimal unsigned zerofill", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "decimal(10,0) unsigned zerofill", mysqlType)
 	javaType, err = getJavaSQLType("2333", columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -615,17 +616,17 @@ func TestGetMySQLType4TimeTypes(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "time", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "time", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "time", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "time(3)", mysqlType)
 	javaType, err := getJavaSQLType("02:20:20", columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -640,17 +641,17 @@ func TestGetMySQLType4TimeTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "datetime", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "datetime", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "datetime", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "datetime(3)", mysqlType)
 	javaType, err = getJavaSQLType("2020-02-20 02:20:20", columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -665,17 +666,17 @@ func TestGetMySQLType4TimeTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "timestamp", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "timestamp", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "timestamp", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "timestamp(3)", mysqlType)
 	javaType, err = getJavaSQLType("2020-02-20 02:20:20", columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -689,9 +690,9 @@ func TestGetMySQLType4TimeTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "date", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "date", mysqlType)
 	javaType, err = getJavaSQLType("2020-02-20", columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -706,17 +707,17 @@ func TestGetMySQLType4TimeTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "year", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "year(4)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "year", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "year(4)", mysqlType)
 	javaType, err = getJavaSQLType("2020", columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -736,17 +737,17 @@ func TestGetMySQLType4Char(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "char", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "char(1)", mysqlType)
 
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "char", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "char(123)", mysqlType)
 	javaType, err := getJavaSQLType([]uint8("测试char"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -761,9 +762,9 @@ func TestGetMySQLType4Char(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "varchar", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "varchar(123)", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试varchar"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -783,9 +784,9 @@ func TestGetMySQLType4TextTypes(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "text", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "text", mysqlType)
 	javaType, err := getJavaSQLType([]uint8("测试text"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -794,9 +795,9 @@ func TestGetMySQLType4TextTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinytext", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinytext", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试tinytext"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -805,9 +806,9 @@ func TestGetMySQLType4TextTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "mediumtext", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "mediumtext", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试mediumtext"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -816,9 +817,9 @@ func TestGetMySQLType4TextTypes(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "longtext", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "longtext", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试longtext"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -838,9 +839,9 @@ func TestGetMySQLType4BinaryType(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "binary", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "binary(1)", mysqlType)
 	javaType, err := getJavaSQLType([]uint8("测试binary"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -849,9 +850,9 @@ func TestGetMySQLType4BinaryType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "binary", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "binary(10)", mysqlType)
 
 	sql = `create table test.t2(a int primary key, b varbinary(23))`
@@ -863,9 +864,9 @@ func TestGetMySQLType4BinaryType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "varbinary", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "varbinary(23)", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试varbinary"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -885,9 +886,9 @@ func TestGetMySQLType4BlobType(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "blob", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "blob", mysqlType)
 	javaType, err := getJavaSQLType([]uint8("测试blob"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -896,9 +897,9 @@ func TestGetMySQLType4BlobType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[2].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "tinyblob", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "tinyblob", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试tinyblob"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -907,9 +908,9 @@ func TestGetMySQLType4BlobType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[3].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[3].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "mediumblob", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "mediumblob", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试mediumblob"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -918,9 +919,9 @@ func TestGetMySQLType4BlobType(t *testing.T) {
 	columnInfo, ok = tableInfo.GetColumnInfo(colInfos[4].ID)
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[4].ID]
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "longblob", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "longblob", mysqlType)
 	javaType, err = getJavaSQLType([]uint8("测试longblob"), columnInfo.FieldType.GetType(), flag)
 	require.NoError(t, err)
@@ -941,9 +942,9 @@ func TestGetMySQLType4EnumAndSet(t *testing.T) {
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
 
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "enum", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "enum('a','b','c')", mysqlType)
 
 	javaType, err := getJavaSQLType(uint64(1), columnInfo.FieldType.GetType(), flag)
@@ -954,9 +955,9 @@ func TestGetMySQLType4EnumAndSet(t *testing.T) {
 	require.True(t, ok)
 	flag = tableInfo.ColumnsFlag[colInfos[2].ID]
 
-	mysqlType = getMySQLType(columnInfo, false)
+	mysqlType = utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "set", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "set('a','b','c')", mysqlType)
 
 	javaType, err = getJavaSQLType(uint64(2), columnInfo.FieldType.GetType(), flag)
@@ -977,9 +978,9 @@ func TestGetMySQLType4JSON(t *testing.T) {
 	columnInfo, ok := tableInfo.GetColumnInfo(colInfos[1].ID)
 	require.True(t, ok)
 	flag := tableInfo.ColumnsFlag[colInfos[1].ID]
-	mysqlType := getMySQLType(columnInfo, false)
+	mysqlType := utils.GetMySQLType(columnInfo, false)
 	require.Equal(t, "json", mysqlType)
-	mysqlType = getMySQLType(columnInfo, true)
+	mysqlType = utils.GetMySQLType(columnInfo, true)
 	require.Equal(t, "json", mysqlType)
 
 	javaType, err := getJavaSQLType("{\"key1\": \"value1\"}", columnInfo.FieldType.GetType(), flag)
