+++ pkg/sink/codec/utils/field_types.go
@@ -14,6 +14,9 @@
 package utils
 
 import (
+	"strings"
+
+	timodel "github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tiflow/cdc/model"
@@ -31,22 +34,35 @@ func SetUnsigned(ft *types.FieldType) *types.FieldType {
 	return ft
 }
 
-// SetFlag set the specific flag to the ft
-func SetFlag(ft *types.FieldType, flag uint) *types.FieldType {
-	ft.SetFlag(flag)
-	return ft
-}
-
 // SetElems set the elems to the ft
 func SetElems(ft *types.FieldType, elems []string) *types.FieldType {
 	ft.SetElems(elems)
 	return ft
 }
 
-// NewTextFieldType create a new text field type.
-func NewTextFieldType(tp byte) *types.FieldType {
-	ft := types.NewFieldType(tp)
-	ft.SetCollate(mysql.DefaultCollationName)
-	ft.SetCharset(mysql.DefaultCharset)
-	return ft
+// when encoding the canal format, for unsigned mysql type, add `unsigned` keyword.
+// it should have the form `t unsigned`, such as `int unsigned`
+func withUnsigned4MySQLType(mysqlType string, unsigned bool) string {
+	if unsigned && mysqlType != "bit" && mysqlType != "year" {
+		return mysqlType + " unsigned"
+	}
+	return mysqlType
+}
+
+func withZerofill4MySQLType(mysqlType string, zerofill bool) string {
+	if zerofill && !strings.HasPrefix(mysqlType, "year") {
+		return mysqlType + " zerofill"
+	}
+	return mysqlType
+}
+
+// GetMySQLType get the mysql type from column info
+func GetMySQLType(columnInfo *timodel.ColumnInfo, fullType bool) string {
+	if !fullType {
+		result := types.TypeToStr(columnInfo.GetType(), columnInfo.GetCharset())
+		result = withUnsigned4MySQLType(result, mysql.HasUnsignedFlag(columnInfo.GetFlag()))
+		result = withZerofill4MySQLType(result, mysql.HasZerofillFlag(columnInfo.GetFlag()))
+		return result
+	}
+	return columnInfo.GetTypeDesc()
 }
